Split backend singletons out from data_access so that these classes can be tested without DB autoconnects occurring.
