+++ server/npg/porch/endpoints/tasks.py
@@ -25,7 +25,7 @@ from typing import List
 from npg.porch.models.pipeline import Pipeline
 from npg.porch.models.task import Task
 
-from npg.porchdb.data_access import get_DbAccessor
+from main_config import get_DbAccessor
 
 router = APIRouter(
     prefix="/tasks",
