+++ server/npg/porchdb/data_access.py
@@ -18,34 +18,13 @@
 # You should have received a copy of the GNU General Public License along with
 # this program. If not, see <http://www.gnu.org/licenses/>.
 
-import os
 from sqlalchemy import select
 from sqlalchemy.exc import IntegrityError
-from sqlalchemy.orm import contains_eager, sessionmaker, joinedload
-from sqlalchemy.ext.asyncio import AsyncSession, create_async_engine
-from typing import Optional, Dict, List
+from sqlalchemy.orm import contains_eager, joinedload
+from typing import Optional, List
 
 from npg.porchdb.models import Pipeline as DbPipeline, Task as DbTask, Agent, Event, Base
 from npg.porch.models import Task, Pipeline
-
-config = {
-    'DB_URL': os.environ.get('DB_URL'),
-    'DB_SCHEMA': os.environ.get('DB_SCHEMA') if os.environ.get('DB_SCHEMA') else 'npg_porch'
-}
-
-engine = create_async_engine(
-    config['DB_URL'], future=True
-)
-Base.metadata.schema = config['DB_SCHEMA']
-session_factory = sessionmaker(
-    engine, expire_on_commit=False, class_=AsyncSession
-)
-
-async def get_DbAccessor():
-    'Provides a hook for fastapi to Depend on a DB session in each route'
-    async with session_factory() as session:
-        async with session.begin():
-            yield AsyncDbAccessor(session)
 
 
 class AsyncDbAccessor:
