+++ server/tests/fixtures/deploy_db.py
@@ -9,6 +9,7 @@ from npg.porchdb.models import (
 )
 import npg.porchdb.data_access
 from main import app
+import main_config
 
 @pytest.fixture
 def minimum_data():
@@ -81,5 +82,5 @@ async def fastapi_testclient(async_session) -> TestClient:
         def _get_db_override():
             return npg.porchdb.data_access.AsyncDbAccessor(async_session)
 
-        app.dependency_overrides[npg.porchdb.data_access.get_DbAccessor] = _get_db_override
+        app.dependency_overrides[main_config.get_DbAccessor] = _get_db_override
         yield TestClient(app)
